requirements.txt
amqp==5.3.1
anyio==4.8.0
beautifulsoup4==4.13.4
billiard==4.2.1
boto3==1.35.81
botocore==1.35.99
celery==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudscraper==1.2.71
colorama==0.4.6
contourpy==1.3.3
coverage==7.10.7
cycler==0.12.1
Faker==37.11.0
filelock==3.20.0
flower==2.0.1
fonttools==4.60.1
fsspec==2025.9.0
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.13.0
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.9
kombu==5.5.3
MarkupSafe==3.0.3
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
packaging==25.0
pandas==2.3.3
peewee==3.18.2
pillow==11.1.0
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit==3.0.51
psutil==7.1.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==22.5
pytz==2025.1
PyYAML==6.0.3
redis==6.1.0
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.10.4
scipy==1.16.2
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.40
ultralytics-thop==2.0.17
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
pytest-json-report==1.5.0